requirements.txt
chainlit>=0.7.0
groq>=0.4.0
pandas>=2.0.0
pinecone-client>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-pinecone>=0.0.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
